Move test empty into test generate markdown readme test file
